Fixed problem with indexing for advert targeting.

+++ src/main/java/com/zuehlke/pgadmissions/services/lifecycle/helpers/ResourceServiceHelper.java
@@ -1,5 +1,7 @@
 package com.zuehlke.pgadmissions.services.lifecycle.helpers;
 
+import java.util.concurrent.atomic.AtomicBoolean;
+
 import javax.inject.Inject;
 
 import org.springframework.stereotype.Component;
@@ -7,19 +9,27 @@ import org.springframework.stereotype.Component;
 import com.zuehlke.pgadmissions.services.ResourceService;
 
 @Component
-public class ResourceServiceHelper implements PrismServiceHelper {
+public class ResourceServiceHelper extends PrismServiceHelperAbstract {
 
     @Inject
     private ResourceService resourceService;
+    
+    private AtomicBoolean shuttingDown = new AtomicBoolean(false);
 
     @Override
     public void execute() {
-        resourceService.deleteElapsedStudyOptions();
+        deleteElapsedStudyOptions();
     }
 
     @Override
-    public void shutdown() {
-        return;
+    public AtomicBoolean getShuttingDown() {
+        return shuttingDown;
     }
-
+    
+    private void deleteElapsedStudyOptions() {
+        if (!isShuttingDown()) {
+            resourceService.deleteElapsedStudyOptions();
+        }
+    }
+    
 }
